requirements.txt
# sync
websocket-client>=0.57.0
requests>=2.21.0

# async
websockets>=13
aiohttp>=3.8.1
async_timeout>=4.0.3

# encrypted
cryptography>=35.0.0
py3rijndael>=0.3.3

# dev
mypy>=1.13
pre-commit>=3.5
pytest>=8.3.3
pytest-asyncio>=0.24
aioresponses>=0.7.7
